Added/updated property tax tests. Disabled GIBDD tests #29
